INITIAL: Basic functionality
Basic functionality has arrived! Here, we have a canvas to draw on with a whopping 20 different colors and 5 pen sizes! You can draw freeform with your mouse, fill in a large range of space with the fill tool, and erase! Alongside that, you can have lonely conversations with yourself in the chatbox, which doesn't allow for certain characters to be entered or it will yell at you! Requires JavaScript or it will redirect you to a page telling you that you need JavaScript.

In recap:
 - Main webpage
 - Javascript warning page
 - The ability to draw
 - The ability to fill in a space
 - The ability to erase
 - The ability to change the color of the pen/fill
 - The ability to change the size of the pen/fill
 - Chat functionality