Basic multiplayer functionality
There are no turns yet, and there still isn't a game. But here we are, a working multiplayer thing.

Users can create and join rooms where they can draw on the canvas and chat. Up to 10 players can join a room; more can theoretically join, but I only created 10 colors.

Anything drawn, filled, or erased on one's screen will also happen on anyone else connected to the same room as the user. The same can be applied to if anyone resets the entire canvas. If the user is not connected to a room, they cannot join.

A set of almost 1000 words that can appear as prompts has been created, but the prompt system hasn't. This will come later.

Also: a new system has been set up for serving files from the server using ExpressJS.